+++ SSRVPN_Android/lib/services/clash_service.dart
@@ -53,6 +53,10 @@ class ClashService extends ClashServiceBase {
   VoidCallback? onAutoConnect;
 
   String get corePath => _corePath;
+
+  @override
+  Future<VpnTrafficSample?> readTrafficSample() => _readNativeTrafficSample();
+
   bool get coreExists => File(_corePath).existsSync();
   bool get nativeConnectionTransitioning => _nativeConnectionTransitioning;
   int? get nativeSessionGeneration => _nativeSessionGeneration;