+++ SSRVPN_MacOS/test/desktop_home_screen_test.dart
@@ -1498,6 +1498,9 @@ class _HomeFixture {
 }
 
 class _FakeClashService extends ClashService {
+  @override
+  Future<VpnTrafficSample?> readTrafficSample() async => null;
+
   _FakeClashService({
     this.recordBatchLatencyResults = true,
     bool running = false,